services/ai-chat/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
psycopg2-binary==2.9.7
qdrant-client==1.6.9
transformers==4.35.0
tokenizers==0.14.1
numpy==1.24.3
scikit-learn==1.3.0
requests==2.31.0